Add files via upload
For Task 2 during my internship at Codesoft, I developed a Java program to automate the calculation of total marks, average percentage, and grade assignment based on marks entered for multiple subjects. The program prompts the user to input marks for each subject, sums them to compute the total marks, and calculates the average percentage. Based on predefined grading criteria, it assigns a grade (A to F) and displays the results. This project enhanced my skills in user input handling, control structures, and basic mathematical operations in Java.